Split io.py into smaller files under new io/
This change helps modularize the functions within io.py and io_support/. With the wildcard imports in io/__init__.py, existing imports do not have to be modified.

Test files updated to reflect the new way of organizing things. Note that there is no test_metadata.py or test_print.py but the contents already have full coverage based on the Codecov report.
