+++ augur/io/print.py
@@ -1,0 +1,5 @@
+import sys
+
+
+def print_err(*args):
+    print(*args, file=sys.stderr)
